% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic}
\alias{dic}
\title{sentiments words dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14843 rows and 2 columns.}
\usage{
dic
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
sentiments words dictionary
}
\examples{
  dic
}
\keyword{datasets}
