% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnuSentiLexR-package.R
\docType{package}
\name{KnuSentiLexR-package}
\alias{KnuSentiLexR}
\alias{KnuSentiLexR-package}
\title{KnuSentiLexR: Korean Sentiment Dictionary from KnuSentiLex}
\description{
Korean Sentiment Dictionary Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/KnuSentiLexR}
  \item Report bugs at \url{https://github.com/mrchypark/KnuSentiLexR/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
