% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti_score.R
\name{senti_magnitude}
\alias{senti_magnitude}
\title{sentiments magnitude}
\usage{
senti_magnitude(sentences)
}
\arguments{
\item{sentences}{target sentences}
}
\description{
calculate sentiments magnitude using KnuSentiLex sentiments words dictionary.
}
