% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senti_score.R
\name{senti_score}
\alias{senti_score}
\title{sentiments score}
\usage{
senti_score(sentences)
}
\arguments{
\item{sentences}{target sentences}
}
\description{
calculate sentiments scores using KnuSentiLex sentiments words dictionary.
}
